% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birthweight_smoking}
\alias{birthweight_smoking}
\title{Birthweight Smoking Data Set}
\format{The file contains 3,000 observations on the variables described below:
\describe{
\item{birthweight}{birth weight of infant (in grams)}
\item{smoker}{indicator equal to one if the mother smoked during pregnancy and zero, otherwise}
\item{age}{age of Mother}
\item{educ}{years of educational attainment by Mother (more than 16 years coded as 17)}
\item{unmarried}{indicator = 1 if mother is unmarried}
\item{alcohol}{indicator = 1 if mother drank alcohol during pregnancy}
\item{drinks}{number of drinks per week}
\item{tripre1}{indicator = 1 if 1st prenatal care visit in 1st trimester}
\item{tripre2}{indicator = 1 if 1st prenatal care visit in 2nd trimester}
\item{tripre3}{indicator = 1 if 1st prenatal care visit in 3rd trimester}
\item{tripre0}{indicator = 1 if no prenatal visits}
\item{nprevisit}{total number of prenatal visits}
}}
\source{
Stock and Watson's Introduction to Econometrics, 3rd Edition. See \url{https://wps.pearsoned.com/wps/media/objects/11422/11696965/data3eu/Birthweight_Smoking_Description.pdf}
}
\usage{
birthweight_smoking
}
\description{
The datafile Birthweight_Smoking is from the 1989 linked National Natality-Mortality Detail
files, which contains a census of infant births and deaths.
}
\details{
These data were provided by Porfessor Douglas Almond, Kenneth Chay, and David Lee and area
subset of the data used in their paper “The Costs of Low Birth Weight,” Quarterly Journal of
Economics, August 2005, 120(3): 1031-1083.
}
\keyword{datasets}
