% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caschool}
\alias{caschool}
\title{California Test Score Data Set}
\format{The file contains 3,000 observations on the variables described below:
\describe{
\item{dist_code}{district code}
\item{read_scr}{average reading test score}
\item{math_scr}{average math test score}
\item{county}{county}
\item{district}{district}
\item{gr_span}{grade span of district}
\item{enrl_tot}{total enrollment}
\item{teachers}{number of teachers}
\item{computer}{number of computers}
\item{testscr}{average test score, (read_scr + math_scr)/2}
\item{comp_stu}{computers per student, (computer / enrl_tot)}
\item{expn_stu}{expentitures per student ($)}
\item{str}{student teacher ratio (enrl_tot / teachers)}
\item{el_pct}{percent of english learners}
\item{meal_pct}{percent qualifying for reduced-price lunch}
\item{calw_pct}{percent qualifying for calworks}
\item{avginc}{district average income (in $1000's)}
}}
\source{
Stock and Watson's Introduction to Econometrics, 3rd Edition. See \url{https://wps.pearsoned.com/aw_stock_ie_3/178/45691/11696965.cw/index.html}
}
\usage{
caschool
}
\description{
The California Standardized Testing and Reporting (STAR) dataset contains data on test performance,
school characteristics and student demographic backgrounds. The data used here are from all 420 K-6
and K-8 districts in California with data available for 1998 and 1999.
}
\keyword{datasets}
