% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{stata_ci}
\alias{stata_ci}
\title{stata_ci}
\usage{
stata_ci(y, x)
}
\arguments{
\item{y}{\itemize{
\item y variable
}}

\item{x}{\itemize{
\item x variable
}}
}
\value{
a matrix (or vector) with columns giving lower and upper confidence limits for each parameter
}
\description{
Finds the confidence interval of a fitted linear model.
}
\examples{
stata_ci(caschool$testscr, caschool$avginc)
stata_ci(mtcars$cyl, mtcars$mpg)

\dontrun{
stata_ci(testscr, avginc)
}

}
