% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter5_functions.R
\name{stata_desc}
\alias{stata_desc}
\title{stata_desc}
\usage{
stata_desc(dataset)
}
\arguments{
\item{dataset}{\itemize{
\item a data table
}}
}
\value{
list of the dataset attributes including every column of the dataset
}
\description{
Gives a glimpse of the dataset.
Provides detailed information of the dataset including the first 10 values of each of the columns.
}
\examples{
stata_desc(caschool)
stata_desc(mtcars)

\dontrun{
stata_desc(caschool, mtcars)
}

}
