% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{stata_list_10}
\alias{stata_list_10}
\title{stata_list_10}
\usage{
stata_list_10(dataset)
}
\arguments{
\item{dataset}{\itemize{
\item an R object which can be coerced to a data frame
}}
}
\value{
a new window containing the first ten rows of a given dataset
}
\description{
View dataset
Allows user to see a the first ten rows of a given dataset in a new window.
}
\examples{
stata_list_10(caschool)
stata_list_10(mtcars)

\dontrun{
stata_list_10(caschool, mtcars)
}

}
