% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter5_functions.R
\name{stata_list_2}
\alias{stata_list_2}
\title{stata_list_2}
\usage{
stata_list_2(dataset)
}
\arguments{
\item{dataset}{\itemize{
\item an R object which can be coerced to a data frame
}}
}
\value{
the first two observations of a dataset in a table
}
\description{
View dataset
Allows user to see a the first two rows of a given dataset.
}
\examples{
stata_list_2(caschool)
stata_list_2(mtcars)

\dontrun{
stata_list_2(caschool, mtcars)
}

}
