% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_functions.R
\name{stata_logit}
\alias{stata_logit}
\title{stata_logit}
\usage{
stata_logit(y, x)
}
\arguments{
\item{y}{\itemize{
\item y variable
}}

\item{x}{\itemize{
\item x variable
}}
}
\value{
list of the result summary of a fitted logistic model
}
\description{
Produces summary results of a fitted logistic model
}
\examples{
stata_logit(birthweight_smoking$alcohol, birthweight_smoking$birthweight)
stata_logit(birthweight_smoking$alcohol, birthweight_smoking$educ)

\dontrun{
stata_logit(alcohol, birthweight)
stata_logit(alcohol, educ)
}

}
