% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{stata_reg}
\alias{stata_reg}
\title{stata_reg}
\usage{
stata_reg(y, x)
}
\arguments{
\item{y}{- y variable}

\item{x}{- x variale}
}
\value{
list of result summary of a fitted linear model
}
\description{
Produces result summaries of a fitted model.
}
\examples{
stata_reg(caschool$testscr, caschool$avginc)
stata_reg(mtcars$cyl, mtcars$mpg)

\dontrun{
stata_reg(testscr, avginc)
}

}
