% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter5_functions.R
\name{stata_sum_var}
\alias{stata_sum_var}
\title{stata_sum_var}
\usage{
stata_sum_var(x, dataset)
}
\arguments{
\item{x}{\itemize{
\item a numeric vector, variable from a datset
}}

\item{dataset}{\itemize{
\item list, dataset
}}
}
\value{
a vector of statistical summaries
}
\description{
Summary results of a specified variable in a given dataset.
Produces result summaries of the various model fitting functions.
}
\examples{
stata_sum_var(caschool$teachers, caschool)
stata_sum_var(mtcars$cyl, mtcars)

\dontrun{
stata_sum_var(teachers, caschool)
}

}
