% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions.R
\name{stata_logit}
\alias{stata_logit}
\title{stata_logit}
\usage{
stata_logit(y, x)
}
\arguments{
\item{y}{\itemize{
\item y variable
}}

\item{x}{\itemize{
\item x variable
}}
}
\value{
list of the result summary of a fitted logistic model
}
\description{
Produces summary results of a fitted logistic model
}
\examples{
stata_logit()
\dontrun{
stata_logit()
}

}
